desugaring: Split sources into separate files
This patch splits the `desugaring` sources into three separate files.
The visitor implementation now lives within `src/visitor.rs`, the
`src/`lib.rs` file is used to implement the `Desugar` trait from
the pipeline, and `src/desugaring.rs` now contains the actual
logic for desugaring various constructs within the AST.

Additionally, integrate `panic_on_span!` macro.
