+++ compiler/hash-ast-desugaring/Cargo.toml
@@ -9,4 +9,5 @@ rayon = "1.5.0"
 
 hash-ast = {path = "../hash-ast" }
 hash-source = { path = "../hash-source" }
+hash-reporting = {path = "../hash-reporting" }
 hash-pipeline = { path = "../hash-pipeline" }
